test(pty): split live UI suite by scope
